fix: resolve ci lint failures
