+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "(?x)( package-lock\\.json$ |Cargo\\.lock$ |uv\\.lock$ |go\\.sum$ |mcpgateway/sri_hashes\\.json$ )|^\\.secrets\\.baseline$",
     "lines": null
   },
-  "generated_at": "2026-07-28T12:35:16Z",
+  "generated_at": "2026-07-28T15:15:04Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -1208,7 +1208,7 @@
         "hashed_secret": "8d7348c4fe7cb659a9ea4ce55d22de30e3ab2c06",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 59,
+        "line_number": 72,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -1216,7 +1216,7 @@
         "hashed_secret": "252d0a1297cb2a9a2e3caaedd9c233612a1fc78f",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 60,
+        "line_number": 73,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -1224,7 +1224,7 @@
         "hashed_secret": "e689846dfc65621eeab3a906bb8b0ddd52f5c514",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 158,
+        "line_number": 171,
         "type": "Secret Keyword",
         "verified_result": null
       }
