+++ mcpgateway/auth_context.py
@@ -97,7 +97,7 @@ See ``AGENTS.md`` section "Authentication & RBAC Overview" for the full
 policy. The key invariants that this module enforces:
 
 1. ``get_token_teams_from_request`` respects the secure-first semantics of
-   ``auth.normalize_token_teams``: missing ``teams`` claim means public-only,
+   ``normalize_token_teams``: missing ``teams`` claim means public-only,
    not admin bypass.
 2. ``get_rpc_filter_context`` derives ``is_admin`` from the verified JWT
    payload or the trusted internal MCP auth context - NOT from the DB user -
@@ -113,7 +113,6 @@ policy. The key invariants that this module enforces:
 """
 
 # Standard
-import asyncio
 import base64
 from collections.abc import Awaitable, Callable
 from functools import lru_cache
@@ -129,8 +128,8 @@ import orjson
 from sqlalchemy.orm import Session
 
 # First-Party
-from mcpgateway.auth import normalize_token_teams
 from mcpgateway.config import settings
+from mcpgateway.db import EmailUser
 
 # Module-level logger
 logger = logging.getLogger(__name__)
@@ -238,6 +237,51 @@ def _non_uuid_identity(value: Any) -> str | None:
     if not value or _is_uuid_string(value):
         return None
     return value
+
+
+def normalize_token_teams(payload: Dict[str, Any]) -> Optional[List[str]]:
+    """
+    Normalize token teams to a canonical form for consistent security checks.
+
+    SECURITY: This is the single source of truth for token team normalization.
+    All code paths that read token teams should use this function.
+
+    Rules:
+    - "teams" key missing -> [] (public-only, secure default)
+    - "teams" is null + is_admin=true -> None (admin bypass, sees all)
+    - "teams" is null + is_admin=false -> [] (public-only, no bypass for non-admins)
+    - "teams" is [] -> [] (explicit public-only)
+    - "teams" is [...] -> normalized list of string IDs
+
+    Args:
+        payload: The JWT payload dict
+
+    Returns:
+        None for admin bypass, [] for public-only, or list of normalized team ID strings
+    """
+    if "teams" not in payload:
+        return []
+
+    teams = payload.get("teams")
+
+    if teams is None:
+        is_admin = payload.get("is_admin", False)
+        if not is_admin:
+            user_info = payload.get("user", {})
+            is_admin = user_info.get("is_admin", False) if isinstance(user_info, dict) else False
+        if is_admin:
+            return None
+        return []
+
+    normalized: List[str] = []
+    for team in teams:
+        if isinstance(team, dict):
+            team_id = team.get("id")
+            if team_id:
+                normalized.append(str(team_id))
+        elif isinstance(team, str):
+            normalized.append(team)
+    return normalized
 
 
 def get_jwt_user_email_from_payload(payload: dict[str, Any]) -> str | None:
@@ -716,7 +760,7 @@ def get_rpc_filter_context(request: Request, user) -> tuple[Optional[str], Optio
         db_user_is_admin = None
         if user_email:
             # First-Party
-            from mcpgateway.db import EmailUser, SessionLocal  # lazy import — avoids cycle
+            from mcpgateway.db import SessionLocal  # pylint: disable=import-outside-toplevel
 
             _db = SessionLocal()
             try:
@@ -943,7 +987,8 @@ async def set_user_context_from_token(request: Request, payload: dict, db: Sessi
 
     Resolves user ID to email and caches on request.state for performance.
     This helper supports the token migration from email-based to user-ID-based
-    tokens by using get_user_email_from_token() which handles both formats.
+    tokens by using signed email metadata first, then resolving UUID subjects
+    through the provided database session.
 
     Args:
         request: FastAPI request object
@@ -974,13 +1019,21 @@ async def set_user_context_from_token(request: Request, payload: dict, db: Sessi
         >>> request.state.user_id  # doctest: +SKIP
         'user@example.com'
     """
-    # First-Party
-    from mcpgateway.auth import _get_user_by_email_sync  # pylint: disable=import-outside-toplevel
-    from mcpgateway.auth import get_user_email_from_token
+    user_email = get_jwt_user_email_from_payload(payload)
+    db_user = None
 
-    user_email = await get_user_email_from_token(payload, db)
+    if user_email is None:
+        subject = payload.get("sub")
+        if isinstance(subject, str):
+            subject = subject.strip()
+            if subject and _is_uuid_string(subject):
+                db_user = db.query(EmailUser).filter(EmailUser.id == subject).first()
+                user_email = db_user.email if db_user else None
+
+    if user_email and db_user is None:
+        db_user = db.query(EmailUser).filter(EmailUser.email == user_email).first()
+
     request.state.user_email = user_email
     request.state.user_id = payload.get("sub")
-    db_user = await asyncio.to_thread(_get_user_by_email_sync, user_email) if user_email else None
     request.state.is_admin = db_user.is_admin if db_user else False
     request.state.auth_provider = payload.get("auth_provider", "local")
