+++ tests/unit/mcpgateway/test_token_migration.py
@@ -119,9 +119,9 @@ async def test_set_user_context_from_token_with_email():
 
     mock_db_user = MagicMock()
     mock_db_user.is_admin = True
+    db.query.return_value.filter.return_value.first.return_value = mock_db_user
 
-    with patch("mcpgateway.auth._get_user_by_email_sync", return_value=mock_db_user):
-        await set_user_context_from_token(request, payload, db)
+    await set_user_context_from_token(request, payload, db)
 
     assert request.state.user_email == "user@example.com"
     assert request.state.user_id == "user@example.com"
@@ -140,13 +140,10 @@ async def test_set_user_context_from_token_with_user_id():
     db = MagicMock()
     mock_user = MagicMock(spec=EmailUser)
     mock_user.email = "user@example.com"
+    mock_user.is_admin = False
     db.query.return_value.filter.return_value.first.return_value = mock_user
 
-    mock_db_user = MagicMock()
-    mock_db_user.is_admin = False
-
-    with patch("mcpgateway.auth._get_user_by_email_sync", return_value=mock_db_user):
-        await set_user_context_from_token(request, payload, db)
+    await set_user_context_from_token(request, payload, db)
 
     assert request.state.user_email == "user@example.com"
     assert request.state.user_id == TEST_UUID
@@ -161,9 +158,9 @@ async def test_set_user_context_from_token_defaults():
     request.state = MagicMock()
     payload = {"sub": "user@example.com"}
     db = MagicMock()
+    db.query.return_value.filter.return_value.first.return_value = None
 
-    with patch("mcpgateway.auth._get_user_by_email_sync", return_value=None):
-        await set_user_context_from_token(request, payload, db)
+    await set_user_context_from_token(request, payload, db)
 
     assert request.state.user_email == "user@example.com"
     assert request.state.user_id == "user@example.com"
